- detect is there duplicated value in write/read? - remove bad macro DEBUG in test files. - add perf counter for wait for io context.

+++ test/thread_local.cc
@@ -16,7 +16,7 @@ class TestThreadLocal {
         printf("alloc success\n");
       }
     }
-    printf("%lu\n", buf);
+    printf("%llu\n", (unsigned long long)buf);
   }
 };
